Added the webinos.js as per the v0.8.0 tag of webinos.
Running webinos applications in a web browser requires the
use of the webinos.js file, created by installing the webinos
platform.  This commit adds a copy of that, as of the v0.8.0
tag of webinos.  If you use this code on a more recent
version of webinos, you will want to update this tag.